# Consistently use `get_requirement` to cache `Requirement` construction