I have been investigating other easy performance improvements when encountering long backtracking times, so I profiled `pip install --dry-run --ignore-installed apache-airflow[all]` where all packages are cached and with https://github.com/pypa/pip/pull/12657 already applied.

<details>
  <summary>Before call graph</summary>

  ![airflow-before-cache](https://github.com/pypa/pip/assets/8070352/814c085e-2686-436f-8c81-6e4c8876e33a)

</details>

For this scenario [ `_parseNoCache`](https://github.com/pypa/pip/blob/f18bebdddbb04372169a617ebdd865974125c196/src/pip/_vendor/pyparsing/core.py#L804) is taking almost 28% of the run time during profiling, partly because it is being called many times by constructing [`Requirement`](https://github.com/pypa/pip/blob/f18bebdddbb04372169a617ebdd865974125c196/src/pip/_vendor/packaging/requirements.py#L87) in a [loop](https://github.com/pypa/pip/blob/f18bebdddbb04372169a617ebdd865974125c196/src/pip/_internal/metadata/importlib/_dists.py#L219)

Caching the construction of  `Requirement` the performance of `apache-airflow[all]` went from 174 seconds to 143 seconds and  `_parseNoCache` reduced to 3% of the run time during profiling:

<details>
  <summary>After call graph</summary>

![airflow-all-cache-requirements](https://github.com/pypa/pip/assets/8070352/19b2dec4-abca-436c-b8a8-c4aace1b9b41)
</details>

Further, testing performance of the highly pathological requirements `"urllib3>2" "boto3<1.20"` total time went from 559 seconds to 117 seconds.

Testing very simple requirements such as `requests` and `pandas` performance appeared to be about 1-3% better, but they were within the margin of error. However I did observe for these requirements all `Requirement`s objects were constructed at least twice, e.g. for `requests` the final `CacheInfo` was `CacheInfo(hits=11, misses=11, maxsize=None, currsize=11)`, and for `pandas` the final one was `CacheInfo(hits=86, misses=86, maxsize=None, currsize=86)`. So it does make sense there would be a small saving for even requirements that don't backtrack.

I set the maxsize to 1024 preemptively for questions about unlimited memory usage, out of the requirements I tested only `apache-airflow[all]` exceeded the cache size, so I tested a few data points:

* 128 maxsize: `CacheInfo(hits=2176, misses=60088, maxsize=128, currsize=128)`
* 256 maxsize: `CacheInfo(hits=2402, misses=59862, maxsize=256, currsize=256)`
* 512: maxsize: `CacheInfo(hits=55408, misses=6856, maxsize=512, currsize=512)`
* 1024 maxsize: `CacheInfo(hits=56087, misses=6127, maxsize=1024, currsize=1024)`
* 2048 maxsize: `CacheInfo(hits=56307, misses=5906, maxsize=2048, currsize=2048)`
* None maxsize: `CacheInfo(hits=59185, misses=3079, maxsize=None, currsize=3079)`

It appears there's some arbitrary point, which depends on the dependency graph, where once the maxsize is over the hit rate goes up significantly. Out of all the graphs I tested, 1024 seemed to be firmly above that number.

For the `apache-airflow[all]`, the time difference between 1024 and None on my machine was 146 seconds vs. 143 seconds.

Happy to take a different approach if someone spots one.