@notatallshaw 2024-04-30
> * Packaging 22.0+ has a hand-rolled parser (replacing `pyparser`) [which is apparently 95% faster.](https://github.com/pypa/pip/issues/11715#issue-1525805458) Will the caching remain beneficial once

Awesome, I will pull that PR branch and reprofile on it and see if this still takes a sginifciant amount of time.

> * There's already a caching wrapper function for Requirement construction in the utilities, let's use it instead of creating another one. The `maxsize` argument should probably be bumped, however. [1](#user-content-fn-1-1c26466ef8725b9af0e72167637cd5b6)

I didn't realize that existed, I can also profile switching all `Requirement` construction to it and checking cache pressure under different scenarios.