@notatallshaw 2024-04-30
> There's already a caching wrapper function for Requirement construction in the utilities, let's use it instead of creating another one

Fixed, made usage consistent across pip codebase, and bumped maxsize, testing cache presure:

 * `apache-airflow[all]`: `CacheInfo(hits=78198, misses=5803, maxsize=2048, currsize=2048)`
 * `"urllib3>2" "boto3<1.20"`: `CacheInfo(hits=786557, misses=1783, maxsize=2048, currsize=1588)`
 * `pandas`: `CacheInfo(hits=91, misses=86, maxsize=2048, currsize=86)`
 * `requests`: `CacheInfo(hits=17, misses=24, maxsize=2048, currsize=20)`

> Packaging 22.0+ has a hand-rolled parser (replacing `pyparser`) [which is apparently 95% faster.](https://github.com/pypa/pip/issues/11715#issue-1525805458) 

I found, that indeed parsing is significantly faster, but due to the `O(n^2)` nature of resolving the performance improvements were still noticable, especially with pathalogical examples.

With https://github.com/pypa/pip/pull/12300 I found `apache-airflow[all]` improved from 141 seconds to 137 seconds, and `"urllib3>2" "boto3<1.20"` improved from 146 seconds to 112 seconds. 

I still saw very small improvements for simple requirements that didn't involve backtracking, but instead of running for dozens of times I had to run hundreds of time for the statistics to approach significance.

So this PR is a big improvement if https://github.com/pypa/pip/pull/12300  doesn't land, but still a significant improvement for pathalogical requirements if https://github.com/pypa/pip/pull/12300 does land.