@pfmoore 2024-05-01
I made the same comment recently on another PR. The `Requirement` object is a huge chunk of our technical debt, unfortunately. One thing that would be nice is to enforce the immutability that we're starting to rely on - but how we do that without a performance cost isn't clear to me (and adding a performance cost to allow us to implement a performance improvement without risk is clearly silly...)