@notatallshaw 2024-05-01
> but how we do that without a performance cost isn't clear to me (and adding a performance cost to allow us to implement a performance improvement without risk is clearly silly...)

I'm not sure the two actions are actions are actually that counter productive. E.g, If `Requirement` became a frozen dataclass it might be more costly to construct by some constant time, But constant time really isn't a problem (unless it's massive), this PR reduces the number of times it has to construct, in cases of backtracking at least, far more than a constant factor.

And likely https://github.com/pypa/pip/pull/12300 reduces the constant time factor of construction significantly more than the constant time factor increase switching to a frozen data class might incur.