@notatallshaw 2024-05-01
> Unfortunately `Requirement` instance attributes are mutable so mayhem is possible if a requirement is ever modified in-place, but it appears that the codebase does not do that. It'll just have to be something to be careful of in future code reviews.

To be clear though, this PR doesn't introduce the assumption that `Requirement` won't be mutated, as it was already being cached already, just not consistently across pip's codebase.

Further `Requirement` has a `__hash__` method, I believe since it was introduced, it is always precarious to mutate an object with a `__hash__` method, as there's significant risk it could be used as the key in a dictionary. Partly, this is the fun of Python's "we're all consenting adults here" philosophy 😄 .