@notatallshaw 2024-05-04
Rebased now https://github.com/pypa/pip/pull/12300 has landed