@pradyunsg 2024-05-15
I'm mildly concerned about the number of caches we've thrown into the resolver context, especially since we have nothing to enforce mutation prevention. While we have control over all the pieces here, so this is theoretically "safe"; we don't have anything to prevent us from mutating these objects.

We should likely look into ensuring we don't mutate these objects if they've been entered into the cache.