@notatallshaw 2024-05-15
Is there a preferred way you have to prevent mutation?

I would be happy to do a follow up PR that turned the objects I've recently touched into data classes, with the appropriate flags set.

The current resolution algorithm is O(n^2), and if caches aren't used then pip can end up calling things O(n^3) or worse. So while this is the way the resolution works, I forsee myself and others finding more places where caches improve the resolution.