@pradyunsg 2024-05-15
> Is there a preferred way you have to prevent mutation?

None at the moment, hence the "we should look into" statement. To be clear, I am not concerned about this specific PR -- I'm on board for that and it's the broader pattern we have that I am concerned about. 😅