@ichard26 2024-06-28
Any objections to landing this for 24.2 once conflicts are fixed? I can look into introducing an immutable requirement wrapper/version of `packaging.requirements.Requirement` afterwards. I started on a patch, but I realized that it would have massive merge conflicts once this landed, so it would be easier to first merge this and then work on paying down the technical debt.