@ichard26 2024-07-06
Of course this has conflicts again, but under the understanding I'll pursue #12806 after this PR (https://github.com/pypa/pip/pull/12663#issuecomment-2195973645). I'm going to schedule this PR for 24.2. Please feel free to remove this PR from the milestone if you object.